day-3

+++ backend/routes/cart.route.js
@@ -1,7 +1,9 @@
 const { Router } = require("express");
 const { CartModel } = require("../models/cart.model");
+const { auth } = require("../middleware/auth.middleware");
 
 const cartRouter = Router();
+cartRouter.use(auth);
 
 cartRouter.get("/", async (req, res) => {
     let {userID} = req.body;
@@ -19,22 +21,36 @@ cartRouter.post("/add", async (req, res) => {
     }
 })
 
-cartRouter.patch("/update/:id",async(req,res)=>{
-    let {id} = req.params
+cartRouter.patch("/update/:id", async (req, res) => {
+    let { id } = req.params;
     try {
-        await CartModel.findByIdAndUpdate(id,req.body);
-        res.status(200).send({"msg":"cart updated successfully"});
+        let data = await CartModel.findById(id);
+        if (data.userID == req.body.userID) {
+            await CartModel.findByIdAndUpdate(id, req.body);
+            res.send({ "msg": "update successfull" });
+        }
+        else {
+            res.status(400).send({ "msg": "you are not authorised for this action" })
+        }
     } catch (error) {
-        res.status(400).send({"msg":error.message});
+        res.status(400).send({ "msg": error.message })
     }
 })
 
 cartRouter.delete("/delete/:id", async (req, res) => {
-    let { id } = req.params
+    let { id } = req.params;
     try {
-        await CartModel.findByIdAndDelete(id);
-        res.status(200).send({ "msg": "cart deleted successfully" });
+        let data = await CartModel.findById(id);
+        if (data.userID == req.body.userID) {
+            await CartModel.findByIdAndDelete(id);
+            res.send({ "msg": "delete successfull" });
+        }
+        else {
+            res.status(400).send({ "msg": "you are not authorised for this action" })
+        }
     } catch (error) {
-        res.status(400).send({ "msg": error.message });
+        res.status(400).send({ "msg": error.message })
     }
 })
+
+module.exports = {cartRouter}
