+++ backend/routes/comment.route.js
@@ -1,7 +1,9 @@
 const { Router } = require("express");
+const { commentAuth } = require("../middleware/commentauth.middleware");
 const { CommentModel } = require("../models/comment.model");
 
 const commentRouter = Router();
+commentRouter.use(commentAuth);
 
 commentRouter.get("/", async (req, res) => {
     let { productID } = req.body;
@@ -18,11 +20,18 @@ commentRouter.post("/add", async (req, res) => {
         res.status(400).send({ "msg": error.message });
     }
 })
+
 commentRouter.patch("/update/:id", async (req, res) => {
     let { id } = req.params;
     try {
-        await CommentModel.findByIdAndUpdate(id, req.body);
-        res.status(200).send({ "msg": "comment updated successfully" });
+        let data = await CommentModel.findById(id);
+        if (data.userID == req.body.userID) {
+            await CommentModel.findByIdAndUpdate(id, req.body);
+            res.send({ "msg": "update successfull" });
+        }
+        else {
+            res.status(400).send({ "msg": "you are not authorised for this action" })
+        }
     } catch (error) {
         res.status(400).send({ "msg": error.message })
     }
@@ -31,8 +40,14 @@ commentRouter.patch("/update/:id", async (req, res) => {
 commentRouter.delete("/delete/:id", async (req, res) => {
     let { id } = req.params;
     try {
-        await CommentModel.findByIdAndDelete(id);
-        res.status(200).send({ "msg": "comment deleted successfully" });
+        let data = await CommentModel.findById(id);
+        if (data.userID == req.body.userID) {
+            await CommentModel.findByIdAndDelete(id);
+            res.send({ "msg": "delete successfull" });
+        }
+        else {
+            res.status(400).send({ "msg": "you are not authorised for this action" })
+        }
     } catch (error) {
         res.status(400).send({ "msg": error.message })
     }
