+++ backend/routes/order.route.js
@@ -1,7 +1,9 @@
 const { Router } = require("express");
 const { OrderModel } = require("../models/order.model");
+const { ordertAuth } = require("../middleware/orderauth.middleware");
 
 const orderRouter = Router();
+orderRouter.use(ordertAuth);
 
 orderRouter.get("/", async (req, res) => {
     let { userID } = req.body;
@@ -20,21 +22,35 @@ orderRouter.post("/add", async (req, res) => {
 })
 
 orderRouter.patch("/update/:id", async (req, res) => {
-    let { id } = req.params
+    let { id } = req.params;
     try {
-        await OrderModel.findByIdAndUpdate(id, req.body);
-        res.status(200).send({ "msg": "order updated successfully" });
+        let data = await OrderModel.findById(id);
+        if (data.userID == req.body.userID) {
+            await OrderModel.findByIdAndUpdate(id, req.body);
+            res.send({ "msg": "update successfull" });
+        }
+        else {
+            res.status(400).send({ "msg": "you are not authorised for this action" })
+        }
     } catch (error) {
-        res.status(400).send({ "msg": error.message });
+        res.status(400).send({ "msg": error.message })
     }
 })
 
 orderRouter.delete("/delete/:id", async (req, res) => {
-    let { id } = req.params
+    let { id } = req.params;
     try {
-        await OrderModel.findByIdAndDelete(id);
-        res.status(200).send({ "msg": "order deleted successfully" });
+        let data = await OrderModel.findById(id);
+        if (data.userID == req.body.userID) {
+            await OrderModel.findByIdAndDelete(id);
+            res.send({ "msg": "delete successfull" });
+        }
+        else {
+            res.status(400).send({ "msg": "you are not authorised for this action" })
+        }
     } catch (error) {
-        res.status(400).send({ "msg": error.message });
+        res.status(400).send({ "msg": error.message })
     }
 })
+
+module.exports = {orderRouter};
